Basic gravity, momentum, and return button
A simple physics function has been implemented to track the position of the image and give it gravity, along with tracking its speed to allow the user to throw it. At the moment, it seems easy to throw the image off-screen and make it inaccessible. A simple fix is adding a "return" button which calls the bottom_right function, returning the image to its initial start location.